+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking
+
+- The function `redirectToRoute()` will now throw a `Folded\Exceptions\RouteNotFoundException` instead of a `OutOfRangeException` in case the route is not found.
+
 ## [0.3.0] 2020-09-15
 
 ### Added